Add "restart training" button
If a training job fails sometimes you just really want to restart it.
