+++ app/src/server/tasks/fineTuning/trainFineTune.task.ts
@@ -10,13 +10,27 @@ import {
 } from "~/modelProviders/fine-tuned/getCompletion";
 import { env } from "~/env.mjs";
 import { trainerv1 } from "~/server/modal-rpc/clients";
+import { trainOpenaiFineTune } from "./trainOpenaiFineTune";
 
 export type TrainFineTuneJob = {
   fineTuneId: string;
 };
 
 export const trainFineTune = defineTask<TrainFineTuneJob>("trainFineTune", async (task) => {
-  const { fineTuneId } = task;
+  const fineTune = await prisma.fineTune.findUnique({
+    where: { id: task.fineTuneId },
+  });
+
+  if (!fineTune) return;
+
+  if (fineTune.baseModel === "GPT_3_5_TURBO") {
+    await trainOpenaiFineTune(fineTune.id);
+  } else {
+    await trainModalFineTune(fineTune.id);
+  }
+});
+
+const trainModalFineTune = async (fineTuneId: string) => {
   const fineTune = await prisma.fineTune.findUnique({
     where: { id: fineTuneId },
     include: {
@@ -90,7 +104,7 @@ export const trainFineTune = defineTask<TrainFineTuneJob>("trainFineTune", async
       },
     });
   }
-});
+};
 
 const formatTrainingRow = (row: TrainingRow, stringsToPrune: string[]) => {
   const instruction = pruneInputMessagesStringified(row.input, stringsToPrune);
